+++ include/nstd/os/windows.h
@@ -1,4 +1,5 @@
 #ifndef NSTD_OS_WINDOWS_H_INCLUDED
 #define NSTD_OS_WINDOWS_H_INCLUDED
 #include "windows/alloc.h"
+#include "windows/io.h"
 #endif