+++ src/os/windows.rs
@@ -1,2 +1,4 @@
 #[cfg(feature = "nstd_os_alloc")]
 pub mod alloc;
+#[cfg(feature = "nstd_os_io")]
+pub mod io;